// frontend/src/api/generated/model/prReviewComment.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */
import type { PrReviewCommentInReplyToId } from './prReviewCommentInReplyToId';
import type { PrReviewCommentLine } from './prReviewCommentLine';
import type { PrReviewCommentOriginalLine } from './prReviewCommentOriginalLine';
import type { PrReviewCommentReactions } from './prReviewCommentReactions';
import type { DiffSide } from './diffSide';
import type { GitHubUser } from './gitHubUser';

export interface PrReviewComment {
  body: string;
  commit_id: string;
  created_at: string;
  html_url: string;
  /** @minimum 0 */
  id: number;
  /** @minimum 0 */
  in_reply_to_id?: PrReviewCommentInReplyToId;
  /** @minimum 0 */
  line?: PrReviewCommentLine;
  /** @minimum 0 */
  original_line?: PrReviewCommentOriginalLine;
  path: string;
  reactions?: PrReviewCommentReactions;
  side: DiffSide;
  updated_at: string;
  user: GitHubUser;
}